[Arch] Extract shared primitives into src/shared/
Move genuinely shared, cross-feature code into src/shared/:
- src/components/ui/ -> src/shared/components/
- src/components/layout/ -> src/shared/layout/
- src/components/brand/ -> src/shared/brand/
- src/components/theme/ -> src/shared/theme/
- src/components/icons/ -> src/shared/icons/
- src/lib/ -> src/shared/lib/

All import paths updated across the codebase. Build passes.

Resolves ROB-14
