+++ content/posts/frontendmentor-development-process.mdx
@@ -53,7 +53,7 @@ import type { Metadata } from "next";
 import { Inter as FontSans } from "next/font/google"
 import "./globals.css";
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shared/lib/utils"
 
 const fontSans = FontSans({
   subsets: ["latin"],
