+++ src/components/content/post-image.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { AspectRatio } from "@/components/ui/aspect-ratio";
+import { AspectRatio } from "@/shared/components/aspect-ratio";
 import Image from "next/image";
 
 interface PostImageProps {
