+++ src/components/playground/controls/button-control.tsx
@@ -1,7 +1,7 @@
 // src/components/playground/controls/ButtonControl.tsx
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 import { ButtonControlProps } from "../types";
