+++ src/components/playground/controls/range-control.tsx
@@ -1,7 +1,7 @@
 // src/components/playground/controls/RangeControl.tsx
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 import { RangeControlProps } from "../types";
