+++ src/components/playground/controls/select-control.tsx
@@ -1,7 +1,7 @@
 // src/components/playground/controls/SelectControl.tsx
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 import { SelectControlProps } from "../types";
