+++ src/components/playground/playground-canvas.tsx
@@ -1,7 +1,7 @@
 // src/components/playground/PlaygroundCanvas.tsx
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 import { PlaygroundCanvasProps } from "./types";
