+++ src/shared/components/aspect-ratio.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 interface AspectRatioProps extends React.HTMLAttributes<HTMLDivElement> {
   ratio?: number;
