+++ src/shared/components/avatar.stories.tsx
