+++ src/shared/components/avatar.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 import { Avatar as AvatarPrimitive } from "@base-ui/react/avatar";
 
 const Avatar = React.forwardRef<
