+++ src/shared/components/tooltip.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 import { Tooltip as BaseTooltip } from "@base-ui/react/tooltip";
 
 interface TooltipProps {
