+++ src/shared/icons/skool-icon.tsx
