+++ src/shared/layout/container.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 interface ContainerProps extends React.ComponentProps<"div"> {
   className?: string;
