+++ src/shared/layout/footer.tsx
@@ -1,6 +1,6 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/components/button";
 import { SOCIAL_LINKS } from "@/constants";
-import { renderSocialIcon } from "@/lib/social-icons";
+import { renderSocialIcon } from "@/shared/lib/social-icons";
 import Link from "next/link";
 
 export default function Footer() {
