+++ src/shared/lib/social-icons.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { SiGithub, SiMentorcruise, SiX } from "@icons-pack/react-simple-icons";
 import { LucideLinkedin } from "lucide-react";
-import SkoolIcon from "@/components/icons/skool-icon";
+import SkoolIcon from "@/shared/icons/skool-icon";
 import { SOCIAL_LINKS } from "@/constants";
 
 type IconName = (typeof SOCIAL_LINKS)[number]["icon"];
